+++ classes/Gems/Screens/Token/Ask/ShowFirstOpenAsk.php
@@ -26,6 +26,16 @@ class ShowFirstOpenAsk extends AskScreenAbstract
     /**
      *
      * @param \Gems_Tracker_Token $token
+     * @return array Added before all other parameters
+     */
+    public function getParameters(\Gems_Tracker_Token $token)
+    {
+        return ['showLastName' => true];
+    }
+
+    /**
+     *
+     * @param \Gems_Tracker_Token $token
      * @return array Of snippets or false to use original
      */
     public function getSnippets(\Gems_Tracker_Token $token)
@@ -39,6 +49,6 @@ class ShowFirstOpenAsk extends AskScreenAbstract
      */
     public function getScreenLabel()
     {
-        return $this->_('Show first open token only');
+        return $this->_('Show first open token only - use lastname');
     }
 }